Structure Refactoring, no change to code was made
