+++ src/app/app.module.ts
@@ -13,19 +13,18 @@ import { NavbarHomeComponent } from './layouts/home/navbar-home/navbar-home.comp
 import { SignupComponent } from './layouts/signup/signup.component';
 import { FormSignupComponent } from './layouts/signup/form-signup/form-signup.component';
 import { CarouselComponent } from './layouts/home/carousel/carousel.component';
-import { ChecklistoneComponent } from './checklists/checklistone/checklistone.component';
-import { ChecklisttwoComponent } from './checklists/checklisttwo/checklisttwo.component';
-import { ChecklistthreeComponent } from './checklists/checklistthree/checklistthree.component';
+import { ChecklistoneComponent } from './checklist/checklistone/checklistone.component';
+import { ChecklisttwoComponent } from './checklist/checklisttwo/checklisttwo.component';
+import { ChecklistthreeComponent } from './checklist/checklistthree/checklistthree.component';
 import { ChecklistModule } from './checklist/checklist.module';
 import { ROUTES } from './app.routes';
 import { SigninComponent } from './layouts/home/signin/signin.component';
 import { FormSigninComponent } from './layouts/home/signin/form-signin/form-signin.component';
 import { SchedulingvisitComponent } from './schedulingvisits/schedulingvisit/schedulingvisit.component';
 import { SchedulingHomeComponent } from './scheduling/scheduling-home/scheduling-home.component';
+import { SchedulingvisitsComponent } from './schedulingvisits/schedulingvisits.component';
 import { SchedulingMeetingModule } from './scheduling/scheduling-meeting/scheduling-meeting.module';
 import { SchedulingMeetingComponent } from './scheduling/scheduling-meeting/scheduling-meeting.component';
-import { SchedulingvisitsComponent } from './schedulingvisits/schedulingvisits.component';
-
 @NgModule({
   declarations: [
     AppComponent,
@@ -42,8 +41,7 @@ import { SchedulingvisitsComponent } from './schedulingvisits/schedulingvisits.c
     SchedulingHomeComponent,
     SigninComponent,
     FormSigninComponent,
-    SchedulingvisitsComponent,
-    SchedulingvisitComponent
+    SchedulingvisitsComponent
   ],
 
   imports: [
