+++ src/app/checklist/checklistone/checklistone.component.css
