+++ src/app/checklist/checklisttwo/checklisttwo.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import {CheckQuestion} from '../checklist.model'
+import {CheckQuestion} from '../checklist.module'
 @Component({
   selector: 'app-checklisttwo',
   templateUrl: './checklisttwo.component.html',
